requirements.txt
gitpython
google-cloud-aiplatform